requirements.txt
streamlit 
python-dotenv
langchain
pyproject-toml
langchain_groq
langchain_community
pillow
transformers
torchvision
torch
torchaudio
soundfile